+++ geopandas/array.py
@@ -13,7 +13,9 @@ from shapely.geometry.base import BaseGeometry
 import shapely.ops
 import shapely.wkt
 
-from ._compat import PANDAS_GE_024, Iterable
+from collections.abc import Iterable
+
+from ._compat import PANDAS_GE_024
 
 
 class GeometryDtype(ExtensionDtype):
@@ -828,6 +830,9 @@ class GeometryArray(ExtensionArray):
         """
         if method is not None:
             raise NotImplementedError("fillna with a method is not yet supported")
+
+        if _isna(value):
+            value = None
         elif not isinstance(value, BaseGeometry):
             raise NotImplementedError(
                 "fillna currently only supports filling with a scalar geometry"
@@ -865,7 +870,12 @@ class GeometryArray(ExtensionArray):
                 return self.copy()
             else:
                 return self
-        return np.array(self, dtype=dtype, copy=copy)
+        elif pd.api.types.is_string_dtype(dtype) and not pd.api.types.is_object_dtype(
+            dtype
+        ):
+            return to_wkt(self).astype(dtype, copy=False)
+        else:
+            return np.array(self, dtype=dtype, copy=copy)
 
     def isna(self):
         """