+++ geopandas/io/tests/test_file.py
@@ -1,5 +1,3 @@
-from __future__ import absolute_import
-
 from collections import OrderedDict
 import datetime
 from distutils.version import LooseVersion